Model/ Repository/ DTO/ Service/ Controller
Model:

Mudança em todos os Models e Repositorys especificamente do tipo long/Long para Integer o que estava dando conflito na integração com o banco de dados.

Repository:

Mudança em todos os Repositorys mas especificamente no tipo do id que antes era Long para Integer pois estava causando conflito na integração com o banco de dados.

DTO:

Criação do package DTO e das classes UsuarioDTO e UsuarioRequestDTO.

Service:

Criado o package Service junto com a classe UsuarioService e que dentro dela há a logica do CRUD.

Controller:

Criado o package Controller junto com a classe UsuarioController onde foi definido os endpoints da entidade Usuario.

+++ back-end/CorpoDeBombeiros/src/main/java/com/pi/CorpoDeBombeiros/model/AnexosModel.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 public class AnexosModel {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID_Anexo")
-    private long idAnexos;
+    private Integer idAnexos;
 
     @Column(name = "Nome_Arquivo", nullable = false)
     private String nomeArquivo;
