+++ back-end/CorpoDeBombeiros/src/main/java/com/pi/CorpoDeBombeiros/repository/AnexosRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnexosRepository extends JpaRepository<AnexosModel, Long> {
+public interface AnexosRepository extends JpaRepository<AnexosModel, Integer> {
 
 }
