+++ back-end/CorpoDeBombeiros/src/main/java/com/pi/CorpoDeBombeiros/repository/SubtipoOcorrenciaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubtipoOcorrenciaRepository extends JpaRepository<SubtipoOcorrenciaModel, Long> {
+public interface SubtipoOcorrenciaRepository extends JpaRepository<SubtipoOcorrenciaModel, Integer> {
 }
