+++ back-end/CorpoDeBombeiros/src/main/java/com/pi/CorpoDeBombeiros/repository/TipoArquivoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoArquivoRepository extends JpaRepository<TipoArquivoModel, Long> {
+public interface TipoArquivoRepository extends JpaRepository<TipoArquivoModel, Integer> {
 }
